+++ pkg/resources/procedure_grant.go
@@ -2,6 +2,7 @@ package resources
 
 import (
 	"errors"
+	"fmt"
 	"strings"
 
 	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/snowflake"
@@ -15,12 +16,6 @@ var validProcedurePrivileges = NewPrivilegeSet(
 )
 
 var procedureGrantSchema = map[string]*schema.Schema{
-	"procedure_name": {
-		Type:        schema.TypeString,
-		Optional:    true,
-		Description: "The name of the procedure on which to grant privileges immediately (only valid if on_future is false).",
-		ForceNew:    true,
-	},
 	"arguments": {
 		Type: schema.TypeList,
 		Elem: &schema.Resource{
@@ -40,23 +35,33 @@ var procedureGrantSchema = map[string]*schema.Schema{
 		Optional:    true,
 		Description: "List of the arguments for the procedure (must be present if procedure has arguments and procedure_name is present)",
 		ForceNew:    true,
+		Deprecated:  "use argument_data_types instead.",
 	},
-	"return_type": {
-		Type:        schema.TypeString,
+	"argument_data_types": {
+		Type:        schema.TypeList,
+		Elem:        &schema.Schema{Type: schema.TypeString},
 		Optional:    true,
-		Description: "The return type of the procedure (must be present if procedure_name is present)",
-		ForceNew:    true,
-	},
-	"schema_name": {
-		Type:        schema.TypeString,
-		Required:    true,
-		Description: "The name of the schema containing the current or future procedures on which to grant privileges.",
+		Description: "List of the argument data types for the procedure (must be present if procedure has arguments and procedure_name is present)",
 		ForceNew:    true,
 	},
 	"database_name": {
 		Type:        schema.TypeString,
 		Required:    true,
 		Description: "The name of the database containing the current or future procedures on which to grant privileges.",
+		ForceNew:    true,
+	},
+	"enable_multiple_grants": {
+		Type:        schema.TypeBool,
+		Optional:    true,
+		Description: "When this is set to true, multiple grants of the same type can be created. This will cause Terraform to not revoke grants applied to roles and objects outside Terraform.",
+		Default:     false,
+		ForceNew:    true,
+	},
+	"on_future": {
+		Type:        schema.TypeBool,
+		Optional:    true,
+		Description: "When this is set to true and a schema_name is provided, apply this grant on all future procedures in the given schema. When this is true and no schema_name is provided apply this grant on all future procedures in the given database. The procedure_name and shares fields must be unset in order to use on_future.",
+		Default:     false,
 		ForceNew:    true,
 	},
 	"privilege": {
@@ -67,11 +72,30 @@ var procedureGrantSchema = map[string]*schema.Schema{
 		ValidateFunc: validation.StringInSlice(validProcedurePrivileges.ToList(), true),
 		ForceNew:     true,
 	},
+	"procedure_name": {
+		Type:        schema.TypeString,
+		Optional:    true,
+		Description: "The name of the procedure on which to grant privileges immediately (only valid if on_future is false).",
+		ForceNew:    true,
+	},
+	"return_type": {
+		Type:        schema.TypeString,
+		Optional:    true,
+		Description: "The return type of the procedure (must be present if procedure_name is present)",
+		ForceNew:    true,
+		Deprecated:  "return_type is no longer required. It will be removed in a future release.",
+	},
 	"roles": {
 		Type:        schema.TypeSet,
+		Required:    true,
 		Elem:        &schema.Schema{Type: schema.TypeString},
-		Optional:    true,
 		Description: "Grants privilege to these roles.",
+	},
+	"schema_name": {
+		Type:        schema.TypeString,
+		Optional:    true,
+		Description: "The name of the schema containing the current or future procedures on which to grant privileges.",
+		ForceNew:    true,
 	},
 	"shares": {
 		Type:        schema.TypeSet,
@@ -79,24 +103,10 @@ var procedureGrantSchema = map[string]*schema.Schema{
 		Optional:    true,
 		Description: "Grants privilege to these shares (only valid if on_future is false).",
 	},
-	"on_future": {
-		Type:        schema.TypeBool,
-		Optional:    true,
-		Description: "When this is set to true and a schema_name is provided, apply this grant on all future procedures in the given schema. When this is true and no schema_name is provided apply this grant on all future procedures in the given database. The procedure_name and shares fields must be unset in order to use on_future.",
-		Default:     false,
-		ForceNew:    true,
-	},
 	"with_grant_option": {
 		Type:        schema.TypeBool,
 		Optional:    true,
 		Description: "When this is set to true, allows the recipient role to grant the privileges to other roles.",
-		Default:     false,
-		ForceNew:    true,
-	},
-	"enable_multiple_grants": {
-		Type:        schema.TypeBool,
-		Optional:    true,
-		Description: "When this is set to true, multiple grants of the same type can be created. This will cause Terraform to not revoke grants applied to roles and objects outside Terraform.",
 		Default:     false,
 		ForceNew:    true,
 	},
@@ -122,85 +132,82 @@ func ProcedureGrant() *TerraformGrantResource {
 
 // CreateProcedureGrant implements schema.CreateFunc.
 func CreateProcedureGrant(d *schema.ResourceData, meta interface{}) error {
-	var (
-		procedureName      string
-		arguments          []interface{}
-		returnType         string
-		procedureSignature string
-		argumentTypes      []string
-	)
+	var procedureName string
 	if name, ok := d.GetOk("procedure_name"); ok {
 		procedureName = name.(string)
-		if ret, ok := d.GetOk("return_type"); ok {
-			returnType = strings.ToUpper(ret.(string))
-		} else {
-			return errors.New("return_type must be set when specifying procedure_name")
-		}
 	}
 	dbName := d.Get("database_name").(string)
 	schemaName := d.Get("schema_name").(string)
 	priv := d.Get("privilege").(string)
-	futureProcedures := d.Get("on_future").(bool)
+	onFuture := d.Get("on_future").(bool)
 	grantOption := d.Get("with_grant_option").(bool)
-	arguments = d.Get("arguments").([]interface{})
+
+	argumentDataTypes := make([]string, 0)
+
+	if v, ok := d.GetOk("arguments"); ok {
+		arguments := v.([]interface{})
+		for _, argument := range arguments {
+			argumentDataTypes = append(argumentDataTypes, argument.(map[string]interface{})["data_type"].(string))
+		}
+	}
+
+	if v, ok := d.GetOk("argument_data_types"); ok {
+		argumentDataTypes = expandStringList(v.([]interface{}))
+	}
+
 	roles := expandStringList(d.Get("roles").(*schema.Set).List())
 
-	if (procedureName == "") && !futureProcedures {
+	if (procedureName == "") && !onFuture {
 		return errors.New("procedure_name must be set unless on_future is true")
 	}
-	if (procedureName != "") && futureProcedures {
+	if (procedureName != "") && onFuture {
 		return errors.New("procedure_name must be empty if on_future is true")
 	}
-
-	if procedureName != "" {
-		procedureSignature, _, argumentTypes = formatCallableObjectName(procedureName, returnType, arguments)
-	} else {
-		argumentTypes = make([]string, 0)
+	if (schemaName == "") && !onFuture {
+		return errors.New("schema_name must be set unless on_future is true")
 	}
 
 	var builder snowflake.GrantBuilder
-	if futureProcedures {
+	if onFuture {
 		builder = snowflake.FutureProcedureGrant(dbName, schemaName)
 	} else {
-		builder = snowflake.ProcedureGrant(dbName, schemaName, procedureName, argumentTypes)
+		builder = snowflake.ProcedureGrant(dbName, schemaName, procedureName, argumentDataTypes)
 	}
 
 	if err := createGenericGrant(d, meta, builder); err != nil {
 		return err
 	}
 
+	// If this is a on_futures grant then the procedure name and arguments do not get set. This is only used for refresh purposes.
+	var procedureObjectName string
+	if !onFuture {
+		procedureObjectName = fmt.Sprintf("%s(%s)", procedureName, strings.Join(argumentDataTypes, ", "))
+	}
 	grant := &grantID{
 		ResourceName: dbName,
 		SchemaName:   schemaName,
-		ObjectName:   procedureSignature,
+		ObjectName:   procedureObjectName,
 		Privilege:    priv,
 		GrantOption:  grantOption,
 		Roles:        roles,
 	}
-	dataIDInput, err := grant.String()
+	grantID, err := grant.String()
 	if err != nil {
 		return err
 	}
-	d.SetId(dataIDInput)
-
+	d.SetId(grantID)
 	return ReadProcedureGrant(d, meta)
 }
 
 // ReadProcedureGrant implements schema.ReadFunc.
 func ReadProcedureGrant(d *schema.ResourceData, meta interface{}) error {
-	var (
-		procedureName string
-		returnType    string
-		arguments     []interface{}
-		argumentTypes []string
-	)
 	grantID, err := grantIDFromString(d.Id())
 	if err != nil {
 		return err
 	}
 	dbName := grantID.ResourceName
 	schemaName := grantID.SchemaName
-	procedureSignature := grantID.ObjectName
+	procedureObjectName := grantID.ObjectName
 	priv := grantID.Privilege
 
 	if err := d.Set("database_name", dbName); err != nil {
@@ -210,33 +217,24 @@ func ReadProcedureGrant(d *schema.ResourceData, meta interface{}) error {
 	if err := d.Set("schema_name", schemaName); err != nil {
 		return err
 	}
-	futureProceduresEnabled := false
-	if procedureSignature == "" {
-		futureProceduresEnabled = true
+	onFuture := false
+	var procedureName string
+	argumentDataTypes := make([]string, 0)
+	if procedureObjectName == "" {
+		onFuture = true
 	} else {
-		procedureSignatureMap, err := parseCallableObjectName(procedureSignature)
-		if err != nil {
-			return err
-		}
-		procedureName = procedureSignatureMap["callableName"].(string)
-		returnType = procedureSignatureMap["returnType"].(string)
-		arguments = procedureSignatureMap["arguments"].([]interface{})
-		argumentTypes = procedureSignatureMap["argumentTypes"].([]string)
+		procedureName, argumentDataTypes = parseFunctionObjectName(procedureObjectName)
 	}
 
 	if err := d.Set("procedure_name", procedureName); err != nil {
 		return err
 	}
 
-	if err := d.Set("arguments", arguments); err != nil {
+	if err := d.Set("argument_data_types", argumentDataTypes); err != nil {
 		return err
 	}
 
-	if err := d.Set("return_type", returnType); err != nil {
-		return err
-	}
-
-	if err := d.Set("on_future", futureProceduresEnabled); err != nil {
+	if err := d.Set("on_future", onFuture); err != nil {
 		return err
 	}
 
@@ -249,13 +247,13 @@ func ReadProcedureGrant(d *schema.ResourceData, meta interface{}) error {
 	}
 
 	var builder snowflake.GrantBuilder
-	if futureProceduresEnabled {
+	if onFuture {
 		builder = snowflake.FutureProcedureGrant(dbName, schemaName)
 	} else {
-		builder = snowflake.ProcedureGrant(dbName, schemaName, procedureName, argumentTypes)
+		builder = snowflake.ProcedureGrant(dbName, schemaName, procedureName, argumentDataTypes)
 	}
 
-	return readGenericGrant(d, meta, procedureGrantSchema, builder, futureProceduresEnabled, validProcedurePrivileges)
+	return readGenericGrant(d, meta, procedureGrantSchema, builder, onFuture, validProcedurePrivileges)
 }
 
 // DeleteProcedureGrant implements schema.DeleteFunc.
@@ -267,19 +265,15 @@ func DeleteProcedureGrant(d *schema.ResourceData, meta interface{}) error {
 	dbName := grantID.ResourceName
 	schemaName := grantID.SchemaName
 
-	futureProcedures := (grantID.ObjectName == "")
+	procedureObjectName := grantID.ObjectName
+	onFuture := (procedureObjectName == "")
 
 	var builder snowflake.GrantBuilder
-	if futureProcedures {
+	if onFuture {
 		builder = snowflake.FutureProcedureGrant(dbName, schemaName)
 	} else {
-		procedureSignatureMap, err := parseCallableObjectName(grantID.ObjectName)
-		if err != nil {
-			return err
-		}
-		procedureName := procedureSignatureMap["callableName"].(string)
-		argumentTypes := procedureSignatureMap["argumentTypes"].([]string)
-		builder = snowflake.ProcedureGrant(dbName, schemaName, procedureName, argumentTypes)
+		procedureName, argumentDataTypes := parseFunctionObjectName(procedureObjectName)
+		builder = snowflake.ProcedureGrant(dbName, schemaName, procedureName, argumentDataTypes)
 	}
 	return deleteGenericGrant(d, meta, builder)
 }
@@ -309,21 +303,16 @@ func UpdateProcedureGrant(d *schema.ResourceData, meta interface{}) error {
 
 	dbName := grantID.ResourceName
 	schemaName := grantID.SchemaName
-	procedureName := grantID.ObjectName
-	futureProcedures := (procedureName == "")
+	procedureObjectName := grantID.ObjectName
+	onFuture := (procedureObjectName == "")
 
 	// create the builder
 	var builder snowflake.GrantBuilder
-	if futureProcedures {
+	if onFuture {
 		builder = snowflake.FutureProcedureGrant(dbName, schemaName)
 	} else {
-		procedureSignatureMap, err := parseCallableObjectName(grantID.ObjectName)
-		if err != nil {
-			return err
-		}
-		procedureName := procedureSignatureMap["callableName"].(string)
-		argumentTypes := procedureSignatureMap["argumentTypes"].([]string)
-		builder = snowflake.ProcedureGrant(dbName, schemaName, procedureName, argumentTypes)
+		procedureName, argumentDataTypes := parseFunctionObjectName(procedureObjectName)
+		builder = snowflake.ProcedureGrant(dbName, schemaName, procedureName, argumentDataTypes)
 	}
 
 	// first revoke