+++ pkg/resources/warehouse_acceptance_test.go
@@ -56,7 +56,7 @@ func TestAcc_Warehouse(t *testing.T) {
 				ResourceName:            "snowflake_warehouse.w",
 				ImportState:             true,
 				ImportStateVerify:       true,
-				ImportStateVerifyIgnore: []string{"initially_suspended", "wait_for_provisioning"},
+				ImportStateVerifyIgnore: []string{"initially_suspended", "wait_for_provisioning", "query_acceleration_max_scale_factor"},
 			},
 		},
 	})